Scoring logic
Scoring logic is completed for all combos (Sums, Runs and Sets).
Bonus scoring (24/7 and 24 by 7 tiles) implemented. Double Time
score (tile played on 2x space) completed.

Appears to be a bug in identifying Runs where there is a run in
a line but the tile played is not part of the run. For example,
the line (3, 1, 2, 3, 4, 5) where the first 3 was the tile played
scored a run of 6 instead of no combos.