Updates play tile event handling to pure JS
Using dojo to attach the click handler for onPlayTile was causing
multiple listeners on the same space because the listeners were not
removed when the playableSpace class was removed at the end of a turn.

The new implementation eliminates the dojo.connect call and uses
standard JS to add an event listner to the board at setup and the
handler (onPlayTile) checks whether the target has the playableSpace
class or not.

Eliminates multiple listeners and uses delegation at the board level.